authorization/requirements.txt
flask==2.2.2
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
pbkdf2-reboot==1.0
flask_sqlalchemy==3.0.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
redis==4.3.5
gevent==22.10.2
flasgger==0.9.5